+++ mcapdecode/mcapdecode/tests/reader.rs
@@ -11,12 +11,12 @@ use std::{
 #[cfg(feature = "arrow")]
 use arrow::array::Int64Array;
 use mcap::{WriteOptions, Writer, records::MessageHeader};
-#[cfg(feature = "arrow")]
-use mcapdecode::RecordBatchOptions;
 use mcapdecode::{
     McapReader, McapReaderError, PreparedTopic, ReadOptions, TimeRange, TopicDecodeStatus,
     TopicInfo,
 };
+#[cfg(feature = "arrow")]
+use mcapdecode::{McapReaderArrowExt as _, RecordBatchOptions};
 use mcapdecode_core::{
     DataTypeDef, DecoderError, EncodingKey, FieldDef, FieldDefs, MessageDecoder, MessageEncoding,
     SchemaEncoding, TopicDecoder, Value,